Initial Commit
This is the initial commit of the KLAX game. The game works for a single
brick on the paddle and allows loading, unloading and moving of the
paddle.

Also, the game terminates whenever a game over condition arises.

The score, misses and time is displayed at 3 corners.

Also included is GNU GPLv3 license.